ajout factory matrice
