+++ tempProject/app/app.js
@@ -5,10 +5,12 @@ import {HelloIonicPage} from './pages/hello-ionic/hello-ionic';
 import {ListPage} from './pages/list/list';
 import {EquationDegre2} from './pages/equation-degre-2/equation-degre-2';
 import {Matrices} from './pages/matrices/matrices';
+import {MatricesFactory} from './factories/matricesFactory';
 
 
 @App({
   templateUrl: 'build/app.html',
+  providers: [MatricesFactory],
   config: {} // http://ionicframework.com/docs/v2/api/config/Config/
 })
 class MyApp {
